verbs: Move EQ functionality into separate file
This is a pre-cursor to added extended verbs functionality.
