+++ Makefile.am
@@ -113,7 +113,8 @@ if HAVE_VERBS
 _verbs_files = \
 	prov/verbs/src/fi_verbs.h \
 	prov/verbs/src/fi_verbs.c \
-	prov/verbs/src/verbs_cq.c
+	prov/verbs/src/verbs_cq.c \
+	prov/verbs/src/verbs_eq.c
 
 if HAVE_VERBS_DL
 pkglib_LTLIBRARIES += libverbs-fi.la
